refactor: consolidate duplicate sip_isdeleted_safe into utils/ only
ui/sip_isdeleted_safe.py was a legacy copy of utils/sip_isdeleted_safe.py.
Redirect all imports (edit_3d_logic, edit_actions_logic, ui/__init__) to
utils/sip_isdeleted_safe and delete the duplicate file.
